Stack Syntax Learning.java
1. Declare an Empty stack s .
2. Take Single Integer T as input.
3. For next T Lines format

case 1. Print the size of the stack in a separate line.
case 2. Remove an element from the stack. If the stack is empty then print in a separate line.
case 3.Add Integer to the stack s.
case 4. Print an element at the top of the stack. If stack is empty print in a separate line.

Input :

10
3 1
3 2
4
4
2
4
3 4
2
4
1
Output :

2
2
1
1
1